+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/JavaMainWrapper.java
@@ -48,6 +48,7 @@ import org.graalvm.word.WordBase;
 import org.graalvm.word.impl.Word;
 
 import com.oracle.svm.core.jdk.RuntimeSupport;
+import com.oracle.svm.core.jfr.events.ShutdownEvent;
 import com.oracle.svm.core.jni.JNIJavaVMList;
 import com.oracle.svm.core.jni.functions.JNIFunctionTables;
 import com.oracle.svm.core.log.Log;
@@ -193,6 +194,7 @@ public class JavaMainWrapper {
 
         /* Wait for all non-daemon threads to exit. */
         PlatformThreads.singleton().joinAllNonDaemonsInNative();
+        ShutdownEvent.emit("No remaining non-daemon Java threads", false);
 
         try {
             /*