+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/jdk/Target_java_lang_Shutdown.java
@@ -29,6 +29,7 @@ import com.oracle.svm.core.annotate.RecomputeFieldValue;
 import com.oracle.svm.core.annotate.Substitute;
 import com.oracle.svm.core.annotate.TargetClass;
 import com.oracle.svm.core.annotate.TargetElement;
+import com.oracle.svm.core.jfr.events.ShutdownEvent;
 import com.oracle.svm.shared.util.VMError;
 
 @TargetClass(className = "java.lang.Shutdown")
@@ -46,7 +47,8 @@ public final class Target_java_lang_Shutdown {
     }
 
     @Substitute
-    static void beforeHalt() {
+    public static void beforeHalt() {
+        ShutdownEvent.emit("Shutdown requested from Java", true);
     }
 
     /**
@@ -73,13 +75,13 @@ public final class Target_java_lang_Shutdown {
      * This substitution makes a few modifications to {@code Shutdown#exit}:
      * <ul>
      * <li>it omits {@code logRuntimeExit} (exit logging is disabled: GR-45418/JDK-8301627).</li>
-     * <li>it omits {@code beforeHalt} (not implemented).</li>
      * <li>it runs teardown hooks after running shutdown hooks and before halting.</li>
      * </ul>
      */
     @Substitute
     static void exit(int status) {
         synchronized (Target_java_lang_Shutdown.class) {
+            beforeHalt();
             runHooks();
             RuntimeSupport.executeTearDownHooks();
             halt(status);